feature from: add error of img size limit
since nodemailer is being used Vercel has
a rate limir of 4MB this means it rejects
the request so backend doesnt even get to
return a proper JSON error it just returns
a 413 Payload Too Large. We add a check for
Vercels limit now. Plans to compress imgs
before sending.